Update tests after move
